pyproject.toml
[project]
name = "new-app"
version = "0.1.0"
description = "Simple static website using Flask and Bootstrap"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-login>=0.6.2",
    "flask-wtf>=1.1.1",
    "email-validator>=2.0.0",
    "python-dotenv>=1.0.0",
    "flask-mail>=0.9.1",
    "werkzeug>=2.3.0",
    "itsdangerous>=2.1.2",
    "flask-migrate>=4.1.0",
    "psycopg2-binary>=2.9.10",
]
